requirements.txt
asgiref==3.2.2
backcall==0.1.0
decorator==4.4.0
Django==2.2.5
django-braces==1.13.0
django-debug-toolbar==2.0
django-extensions==2.2.1
django-widget-tweaks==1.4.5
gunicorn==19.9.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.6
